+++ rasp-admin/src/main/java/com/jrasp/admin/server/vo/CreateConfigVo.java
@@ -67,4 +67,8 @@ public class CreateConfigVo {
      */
     private String xmlBlockContent;
 
+    /**
+     * 日志文件路径
+     */
+    private String logPath;
 }